Refactor out renderer_draw_text
