+++ source/renderer_shared.c
@@ -1,4 +1,7 @@
 #include "renderer.h"
+#include "lut.h"
+
+#include <string.h>
 
 // This file contains code that's either the exact same across platforms,
 // or so similar that it makes sense to put it in the same file with ifdefs.
@@ -116,4 +119,77 @@ int renderer_get_level_section_from_position(const model_t* model, vec3_t positi
         }
     }
     return n_sections; // -1 means no section
+}
+
+void renderer_draw_text(vec2_t pos, const char* text, const int text_type, const int centered, const pixel32_t color) {
+    // Set to text_type == 0 by default. This way the variables are always initialized
+    int font_x = 0;
+    int font_y = 60;
+    int font_src_width = 7;
+    int font_src_height = 9;
+    int font_dst_width = 7;
+    int font_dst_height = 9;
+    int chars_per_row = 36;
+
+    if (text_type == 1) {
+        font_x = 0;
+        font_y = 80;
+        font_src_width = 16;
+        font_src_height = 16;
+        font_dst_width = 16;
+        font_dst_height = 16;
+        chars_per_row = 16;
+    }
+    else if (text_type == 2) {
+        font_x = 0;
+        font_y = 60;
+        font_src_width = 7;
+        font_src_height = 9;
+        font_dst_width = 14;
+        font_dst_height = 18;
+        chars_per_row = 16;
+    }
+
+    if (centered) {
+        pos.x -= ((strlen(text)) * font_dst_width - (font_dst_width / 2)) * ONE / 2;
+    }
+
+    vec2_t start_pos = pos;
+
+    // Draw each character
+    while (*text) {
+        // Handle special cases
+        if (*text == '\n') {
+            pos.x = start_pos.x;
+            pos.y += font_dst_height * ONE;
+            goto end;
+        }
+        
+        if (*text == '\r') {
+            pos.x = start_pos.x;
+            goto end;
+        }
+
+        if (*text == '\t') {
+            // Get X coordinate relative to start, and round the position up to the nearest multiple of 4
+            scalar_t rel_x = pos.x - start_pos.x;
+            int n_spaces = 4 - ((rel_x / font_dst_width) % 4);
+            pos.x += n_spaces * font_dst_width * ONE;
+            goto end;
+        }
+
+        // Get index in bitmap
+        int index_to_draw = (int)lut_font_letters[(size_t)*text];
+        
+        // Get UV coordinates
+        vec2_t top_left;
+        top_left.x = (font_x + (font_src_width * (index_to_draw % chars_per_row))) * ONE;
+        top_left.y = (font_y + ((index_to_draw / chars_per_row) * font_src_height)) * ONE;
+        vec2_t bottom_right = vec2_add(top_left, (vec2_t){(font_src_width-1)*ONE, (font_src_height-1)*ONE});
+
+        renderer_draw_2d_quad_axis_aligned(pos, (vec2_t){(font_dst_width-1)*ONE, (font_dst_height-1)*ONE}, top_left, bottom_right, (pixel32_t){color.r/2, color.g/2, color.b/2, 255}, 1, 5, 1);
+        pos.x += font_dst_width * ONE;
+        end:
+        ++text;
+    }
 }
